remove abstract models
