--- api/models/__init__.py
@@ -1,3 +1,0 @@
-from .biosample_characteristics import BiosampleCharacteristics
-from .provenance import Provenance
-from .experimental_design import ExperimentalDesign
