+++ api/models/search.py
@@ -1,11 +1,11 @@
 from typing import List
-from pydantic import BaseModel
 
 from api.common.enums import RecordType
+from api.models.base_row_models import RowModel
 
 from .base_response_models import PagedResponseModel
 
-class RecordSearchResult(BaseModel):
+class RecordSearchResult(RowModel):
     id: str # primary_key (identifier) for the record (e.g., ensembl ID)
     description: str # descriptive text
     display: str # display id (e.g. gene symbol)
