+++ api/routers/genomics/queries/gene.py
@@ -1,5 +1,5 @@
 
-from api.models.base_models import QueryDefinition
+from api.models.response_model_properties import QueryDefinition
 
 class GeneQuery(QueryDefinition):
     pass
